motionOrientManager: match three functions
sendStateMail: the three zero stores into the transposed matrix are one
chained assignment, which puts the value in a float register and lets
one store sink into the call's delay slot. SetMotionRequest: its third
parameter is a 32-byte record by value, whose callee copy is the
ldl/ldr block the listing attributes to the declaration line; the six
callers spell the argument that way. shiftMotionOrientBeginFunc: a
block-scoped matrix whose slot a later inline reuses, the attested
16-byte vector union as a store barrier, GetTableArcTan2 returning
short, and an unsigned range test for the two excluded kinds. The
request record and the corrected prototypes live in the headers; every
affected TU keeps its bytes. One jump table carved at 0x620460.

tableSin.h declares GetTableArcTan2 as returning short, which the
callers' bytes need, while the definition in tableSin.c returns int,
which its own bytes need; the definition's TU therefore does not
include the header and repeats the declaration order locally, since
that order fixes the inline tail.

+++ ico2/sugipon/include/tableSin.h
@@ -20,7 +20,7 @@ float GetTableCos(short a0);
 void InitTableSin(void);
 int GetTableArcSin(float x);
 int GetTableArcCos(float x);
-int GetTableArcTan2(float f12, float f13);
+short GetTableArcTan2(float f12, float f13);
 
 /* The declarations below lead this header because their order is load-bearing:
  * gcc 2.9 emits the deferred out-of-line copy of a plain-inline function in
